requirements.txt
fastapi
uvicorn
python-dotenv
kafka-python
psycopg2
pandas
